Move stack generation to the UT #35
- refactoring ModelCompilerAPI - moved interfaces to separate files
- unimportant changes.
